wopr-terminal/requirements.txt
anthropic>=0.18.0
openai>=1.0.0
elevenlabs>=0.2.0
blessed>=1.20.0
pygame>=2.5.0
asyncio
pydantic>=2.0
python-dotenv>=1.0.0
colorama>=0.4.6
